# generation/example_output/coffee_shop_client/models/OrderItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Coffee Shop API
 * API for managing orders and menu in a coffee shop.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@coffeshop.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface OrderItem
 */
export interface OrderItem {
    /**
     * Menu item ID
     * @type {number}
     * @memberof OrderItem
     */
    menuItemId?: number;
    /**
     * Quantity
     * @type {number}
     * @memberof OrderItem
     */
    quantity?: number;
}

/**
 * Check if a given object implements the OrderItem interface.
 */
export function instanceOfOrderItem(value: object): value is OrderItem {
    return true;
}

export function OrderItemFromJSON(json: any): OrderItem {
    return OrderItemFromJSONTyped(json, false);
}

export function OrderItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): OrderItem {
    if (json == null) {
        return json;
    }
    return {
        
        'menuItemId': json['menu_item_id'] == null ? undefined : json['menu_item_id'],
        'quantity': json['quantity'] == null ? undefined : json['quantity'],
    };
}

export function OrderItemToJSON(json: any): OrderItem {
    return OrderItemToJSONTyped(json, false);
}

export function OrderItemToJSONTyped(value?: OrderItem | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'menu_item_id': value['menuItemId'],
        'quantity': value['quantity'],
    };
}